Creating the `Traps` class
